% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{ga_artifact_list}
\alias{ga_artifact_list}
\alias{ga_artifact_table}
\alias{ga_artifact}
\alias{ga_artifact_delete}
\alias{ga_artifact_download}
\title{GitHub Artifacts}
\usage{
ga_artifact_list(owner, repo = NULL, page = NULL, per_page = NULL, ...)

ga_artifact_table(...)

ga_artifact(owner, repo = NULL, artifact_id, ...)

ga_artifact_delete(owner, repo = NULL, artifact_id, ...)

ga_artifact_download(
  owner,
  repo = NULL,
  artifact_id,
  archive_format = "zip",
  ...
)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{page}{page to query.  If \code{NULL}, then will iterate through all pages}

\item{per_page}{number of results per page, max is 100.}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}

\item{artifact_id}{identifier of artifact}

\item{archive_format}{format of archive}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub Artifacts
}
\examples{
a = ga_artifact_list("muschellij2", "pycwa")
tab = ga_artifact_table("muschellij2/pycwa")
a
a_id = a$artifacts[[1]]$id
\donttest{
art = ga_artifact("muschellij2", "pycwa", a$artifacts[[1]]$id)
dl = ga_artifact_download("muschellij2", "pycwa", a$artifacts[[1]]$id)
}
}
