% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{ga_job}
\alias{ga_job}
\alias{ga_job_logs}
\title{GitHub jobs}
\usage{
ga_job(owner, repo, job_id, ...)

ga_job_logs(owner, repo, job_id, ...)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{job_id}{identifier of job}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub jobs
}
