% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_log.R
\name{ga_run_log_config}
\alias{ga_run_log_config}
\title{Extract Log Configuration}
\usage{
ga_run_log_config(zipfile)
}
\arguments{
\item{zipfile}{Log zip file from \link{ga_run_download_log}}
}
\value{
A list of character vectors
}
\description{
Extract Log Configuration
}
\examples{
zipfile = ga_run_download_log("r-lib", "actions",  "551409765")
config = ga_run_log_config(zipfile)
}
