% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{ga_run_list}
\alias{ga_run_list}
\alias{ga_run}
\alias{ga_run_delete}
\alias{ga_run_cancel}
\alias{ga_run_download_log}
\alias{ga_run_delete_log}
\alias{ga_run_rerun}
\alias{ga_run_usage}
\alias{ga_run_jobs}
\alias{ga_run_artifacts}
\title{GitHub runs}
\usage{
ga_run_list(owner, repo, ...)

ga_run(owner, repo, run_id, ...)

ga_run_delete(owner, repo, run_id, ...)

ga_run_cancel(owner, repo, run_id, ...)

ga_run_download_log(owner, repo, run_id, ...)

ga_run_delete_log(owner, repo, run_id, ...)

ga_run_rerun(owner, repo, run_id, ...)

ga_run_usage(owner, repo, run_id, ...)

ga_run_jobs(owner, repo, run_id, page = NULL, per_page = NULL, ...)

ga_run_artifacts(owner, repo, run_id, page = NULL, per_page = NULL, ...)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}

\item{run_id}{identifier of run}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub runs
}
\examples{
runs = ga_run_list("muschellij2", "pycwa")
run_id = runs$workflow_runs[[1]]$id
\dontrun{
run = ga_run("muschellij2", "pycwa", run_id)
run_jobs = ga_run_jobs("muschellij2", "pycwa", run_id)
run_log = ga_run_download_log("muschellij2", "pycwa", run_id)
usage = ga_run_usage("muschellij2", "pycwa", run_id)
run_artifacts = ga_run_artifacts("muschellij2", "pycwa", run_id)
}
}
