% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{ga_secrets}
\alias{ga_secrets}
\alias{ga_secret}
\alias{ga_secret_delete}
\alias{ga_secret_public_key}
\alias{ga_secret_update}
\alias{ga_secret_create}
\alias{ga_org_secrets}
\alias{ga_org_secret_public_key}
\alias{ga_org_secret}
\alias{ga_org_secret_update}
\alias{ga_org_secret_create}
\title{GitHub Actions Secrets}
\usage{
ga_secrets(owner, repo = NULL, page = NULL, per_page = NULL, ...)

ga_secret(owner, repo = NULL, secret_name, ...)

ga_secret_delete(owner, repo = NULL, secret_name, ...)

ga_secret_public_key(owner, repo = NULL, ...)

ga_secret_update(owner, repo = NULL, secret_name, secret, ...)

ga_secret_create(owner, repo = NULL, secret_name, secret, ...)

ga_org_secrets(org, page = NULL, per_page = NULL, ...)

ga_org_secret_public_key(org, ...)

ga_org_secret(org, secret_name, ...)

ga_org_secret_update(
  org,
  secret_name,
  secret,
  visibility = c("all", "private", "selected"),
  selected_repository_ids = NULL,
  ...
)

ga_org_secret_create(
  org,
  secret_name,
  secret,
  visibility = c("all", "private", "selected"),
  selected_repository_ids = NULL,
  ...
)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{page}{page to query.  If \code{NULL}, then will iterate through all pages}

\item{per_page}{number of results per page.}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}

\item{secret_name}{name of the secret you are updating or creating}

\item{secret}{value of the secret that you are updating}

\item{org}{organization identifier}

\item{visibility}{for keys, what visibility should it have:
\code{all} - All repositories in an organization can access the secret.
\code{private} - Private repositories in an organization can access the secret.
\code{selected} - Only specific repositories can access the secret.}

\item{selected_repository_ids}{identifiers of repositories if visibility
is \code{selected}}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub Actions Secrets
}
\examples{
\dontrun{
  have_token = length(gh::gh_token()) > 0
  if (have_token) {
    secrets = ga_secrets("muschellij2/pycwa")
    secrets$secrets[[1]]
    # ga_secret_public_key("muschellij2/pycwa")
    ga_secret_update("muschellij2/pycwa", secret_name = "TEST",
    secret = as.character(Sys.time()))
    ga_secret("muschellij2/pycwa", secret_name = "TEST")
  }
}
}
