% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{ga_workflow_list}
\alias{ga_workflow_list}
\alias{ga_workflow_runs}
\alias{ga_workflow}
\alias{ga_workflow_usage}
\alias{ga_workflow_disable}
\alias{ga_workflow_enable}
\title{GitHub Workflows}
\usage{
ga_workflow_list(owner, repo, ...)

ga_workflow_runs(owner, repo, workflow_id, ...)

ga_workflow(owner, repo, workflow_id, ...)

ga_workflow_usage(owner, repo, workflow_id, ...)

ga_workflow_disable(owner, repo, workflow_id, ...)

ga_workflow_enable(owner, repo, workflow_id, ...)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}

\item{workflow_id}{identifier of workflow}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub Workflows
}
\examples{
w = ga_workflow_list("muschellij2", "pycwa")
workflow_id = w$workflows[[1]]$id
\dontrun{
runs = ga_workflow_runs("muschellij2", "pycwa", workflow_id)
flow = ga_workflow("muschellij2", "pycwa", workflow_id)
usage = ga_workflow_usage("muschellij2", "pycwa", workflow_id)
}
}
