% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{ga_job}
\alias{ga_job}
\alias{ga_job_logs}
\title{GitHub jobs}
\usage{
ga_job(owner, repo = NULL, job_id, ...)

ga_job_logs(owner, repo = NULL, job_id, ...)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{job_id}{identifier of job}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub jobs
}
\examples{
have_token = length(gh::gh_token()) > 0
if (have_token) {
  job_id = "1475662285"
  job_out = ga_job("muschellij2", "pycwa", job_id)
  job_out2 = ga_job("muschellij2/pycwa", job_id = job_id)
  if (difftime(Sys.time(), as.POSIXct(job_out$completed_at), "days")<= 90) {
    job_log = ga_job_logs("muschellij2", "pycwa", job_id)
    job_log = ga_job_logs("muschellij2/pycwa", job_id = job_id)
  }
}
}
