% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_log.R
\name{ga_run_log_config}
\alias{ga_run_log_config}
\title{Extract Log Configuration}
\usage{
ga_run_log_config(zipfile, make_data_frame = FALSE)
}
\arguments{
\item{zipfile}{Log zip file from \link{ga_run_download_log}}

\item{make_data_frame}{Make it into a \code{data.frame}}
}
\value{
A list of character vectors
}
\description{
Extract Log Configuration
}
\examples{
zipfile = ga_run_download_log("r-lib", "actions",  "551409765")
config = ga_run_log_config(zipfile)
zipfile = ga_run_download_log("muschellij2", "pycwa",  "392215958")

}
