% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{ga_workflow_list}
\alias{ga_workflow_list}
\alias{ga_workflow_runs}
\alias{ga_workflow_runs_table}
\alias{ga_workflow}
\alias{ga_workflow_usage}
\alias{ga_workflow_disable}
\alias{ga_workflow_enable}
\alias{ga_workflow_dispatch}
\title{GitHub Workflows}
\usage{
ga_workflow_list(owner, repo = NULL, page = NULL, per_page = NULL, ...)

ga_workflow_runs(
  owner,
  repo = NULL,
  workflow_id,
  page = NULL,
  per_page = NULL,
  ...
)

ga_workflow_runs_table(...)

ga_workflow(owner, repo = NULL, workflow_id, ...)

ga_workflow_usage(owner, repo = NULL, workflow_id, ...)

ga_workflow_disable(owner, repo = NULL, workflow_id, ...)

ga_workflow_enable(owner, repo = NULL, workflow_id, ...)

ga_workflow_dispatch(owner, repo = NULL, workflow_id, ref, ...)
}
\arguments{
\item{owner}{owner of repo/username}

\item{repo}{repository name}

\item{page}{page to query.  If \code{NULL}, then will iterate through all pages}

\item{per_page}{number of results per page.}

\item{...}{additional arguments to pass to \code{\link[gh:gh]{gh::gh()}}}

\item{workflow_id}{identifier of workflow}

\item{ref}{The git reference for the workflow. The reference can be a branch or tag name.}
}
\value{
Answer from the API as a \code{ga_response} object, which is also a list.
Failed requests will generate an R error.
Requests that generate a raw response will return a raw vector.
}
\description{
GitHub Workflows
}
\examples{
\dontrun{
have_token = length(gh::gh_token()) > 0
if (have_token) {
  w = ga_workflow_list("muschellij2/pycwa", page = 1)
  w = ga_workflow_list("muschellij2", "pycwa")
  workflow_id = w$workflows[[1]]$id
}

if (have_token) {
  runs = ga_workflow_runs("muschellij2", "pycwa", workflow_id)
  flow = ga_workflow("muschellij2", "pycwa", workflow_id)
  usage = ga_workflow_usage("muschellij2", "pycwa", workflow_id)
}
}
}
